+++ projects/components/src/lib/table/table.spec.ts
@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import { Component, ChangeDetectionStrategy } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 
@@ -9,6 +9,7 @@ import { AuTable } from './table';
 
 @Component({
   imports: [AuTable, AuTableColumn],
+  changeDetection: ChangeDetectionStrategy.OnPush,
   template: `
     <au-table
       [data]="rows"
@@ -60,6 +61,7 @@ class TableHost {
 
 @Component({
   imports: [AuTable, AuTableColumn, AuTableCellDef],
+  changeDetection: ChangeDetectionStrategy.OnPush,
   template: `
     <au-table [data]="rows">
       <au-table-column
@@ -82,6 +84,7 @@ class CustomCellHost {
 
 @Component({
   imports: [AuTable, AuTableColumn],
+  changeDetection: ChangeDetectionStrategy.OnPush,
   template: `
     <au-table
       [data]="rows"